+++ edsnlp/pipes/__init__.py
@@ -123,3 +123,10 @@ if TYPE_CHECKING:
     from .trainable.ner_crf.factory import create_component as ner_crf
     from .trainable.span_classifier.factory import create_component as span_classifier
     from .trainable.span_linker.factory import create_component as span_linker
+    from .trainable.embeddings.span_pooler.factory import create_component as span_pooler
+    from .trainable.embeddings.transformer.factory import create_component as transformer
+    from .trainable.embeddings.text_cnn.factory import create_component as text_cnn
+    from .misc.split import Split as split
+    from .misc.explode import Explode as explode
+    from .llm.llm_markup_extractor.factory import create_component as llm_markup_extractor
+    from .llm.llm_span_qualifier.factory import create_component as llm_span_qualifier